Dev api

+++ services/Turma/DeleteTurma.js
@@ -1,27 +1,33 @@
 const Turma = require('../../models/Turma')
+const Professor = require('../../models/Professor')
+const Aluno = require('../../models/Aluno')
 
 const delete_turma =  async (req,res)=>{
-    const {id} = req.params
-    try{
-        //Tenta encontrar por id 
-        const deleteuser = await Turma.findByIdAndDelete(id)
-
-        //verifica se achou alguma turma pelo id
-        if(deleteuser){
-            res.status(200).json({
-            message : 'Turma deletada com sucesso'})
-        } else {
-        //Caso não encontre turma
-            res.status(404).json({
-            message : 'Turma não encontrada'})  
-            }
-        } catch (error){
-        //Caso tenha algum erro na conexão ou em algo relacionado ao banco
-            res.status(500).json({
-                message : 'Erro ao deletar a Turma',
-                error: error.message
-            })
-        }
+    const {id} = req.params;
+    console.log(id);
+    const turma = await Turma.findById(id);
+    if(!turma){
+        return res.status(404).json({
+            message: 'Turma não encontrada'
+        });
+    }
+    const alunos = await Aluno.find({ turma: id });
+    if(!turma.professor && !alunos) {
+        await Turma.findByIdAndDelete(id)
+        return res.status(200).json({
+            message: 'Turma deletada com sucesso',
+        });
+    }
+    // deleta alunos com a turma
+    alunos.forEach(async (aluno) => {
+        await Aluno.findByIdAndDelete(aluno._id);
+    })
+    // deleta professor e turma
+    await Professor.findByIdAndDelete(turma.professor);
+    await Turma.findByIdAndDelete(id);
+    return res.status(200).json({
+        message: 'Turma deletada com sucesso',
+    });
 }
 
 module.exports = delete_turma
